+++ src/lib/jp2/t1/t1_ht/T1HT.h
@@ -33,7 +33,7 @@ namespace t1_ht {
 
 class T1HT: public T1Interface {
 public:
-	T1HT(bool isEncoder, grk_tcp *tcp, uint16_t maxCblkW, uint16_t maxCblkH);
+	T1HT(bool isEncoder, grk_tcp *tcp, uint32_t maxCblkW, uint32_t maxCblkH);
 	virtual ~T1HT();
 
 	void preEncode(encodeBlockInfo *block, grk_tcd_tile *tile, uint32_t &max);
